# time_attendant/time_attendant_openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/check_user/":{"post":{"summary":"Check User","description":"API สำหรับตรวจสอบผู้ใช้","operationId":"check_user_check_user__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/check_in/":{"post":{"summary":"Check In","description":"API สำหรับบันทึกเวลาเข้างาน","operationId":"check_in_check_in__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/check_out/":{"post":{"summary":"Check Out","description":"API สำหรับบันทึกเวลาออกงาน","operationId":"check_out_check_out__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AttendanceRequest":{"properties":{"username":{"type":"string","title":"Username"}},"type":"object","required":["username"],"title":"AttendanceRequest","description":"โมเดลสำหรับ request body ในการลงเวลา"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserRequest","description":"โมเดลสำหรับ request body"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}